Improved encapsulation, folders arrangement, added Service Collection Extension
