+++ PersonalFinancer.Web/Areas/Identity/Pages/Account/Manage/Email.cshtml.cs
@@ -82,7 +82,7 @@ namespace PersonalFinancer.Web.Areas.Identity.Pages.Account.Manage
 				var callbackUrl = Url.Page(
 					"/Account/ConfirmEmailChange",
 					pageHandler: null,
-					values: new { area = "Identity", userId = userId, email = Input.NewEmail, code = code },
+					values: new { area = "Identity", userId, email = Input.NewEmail, code },
 					protocol: Request.Scheme);
 				await emailSender.SendEmailAsync(
 					Input.NewEmail,
@@ -118,7 +118,7 @@ namespace PersonalFinancer.Web.Areas.Identity.Pages.Account.Manage
 			var callbackUrl = Url.Page(
 				"/Account/ConfirmEmail",
 				pageHandler: null,
-				values: new { area = "Identity", userId = userId, code = code },
+				values: new { area = "Identity", userId, code },
 				protocol: Request.Scheme);
 			await emailSender.SendEmailAsync(
 				email,
