+++ PersonalFinancer.Web/Areas/Identity/Pages/Account/Manage/PersonalData.cshtml.cs
@@ -8,19 +8,16 @@ namespace PersonalFinancer.Web.Areas.Identity.Pages.Account.Manage
 	public class PersonalDataModel : PageModel
     {
         private readonly UserManager<ApplicationUser> userManager;
-        private readonly ILogger<PersonalDataModel> logger;
 
-        public PersonalDataModel(
-            UserManager<ApplicationUser> userManager,
-            ILogger<PersonalDataModel> logger)
+        public PersonalDataModel(UserManager<ApplicationUser> userManager)
         {
             this.userManager = userManager;
-            this.logger = logger;
         }
 
         public async Task<IActionResult> OnGet()
         {
             var user = await userManager.GetUserAsync(User);
+
             if (user == null)
             {
                 return NotFound($"Unable to load user with ID '{userManager.GetUserId(User)}'.");
