+++ PersonalFinancer.Web/Areas/Identity/Pages/Account/Manage/ResetAuthenticator.cshtml.cs
@@ -28,6 +28,7 @@ namespace PersonalFinancer.Web.Areas.Identity.Pages.Account.Manage
 		public async Task<IActionResult> OnGet()
 		{
 			var user = await userManager.GetUserAsync(User);
+
 			if (user == null)
 			{
 				return NotFound($"Unable to load user with ID '{userManager.GetUserId(User)}'.");
@@ -46,7 +47,7 @@ namespace PersonalFinancer.Web.Areas.Identity.Pages.Account.Manage
 
 			await userManager.SetTwoFactorEnabledAsync(user, false);
 			await userManager.ResetAuthenticatorKeyAsync(user);
-			var userId = await userManager.GetUserIdAsync(user);
+
 			logger.LogInformation("User with ID '{UserId}' has reset their authentication app key.", user.Id);
 
 			await signInManager.RefreshSignInAsync(user);
