+++ PersonalFinancer.Web/Areas/Identity/Pages/Account/Manage/TwoFactorAuthentication.cshtml.cs
@@ -10,16 +10,13 @@ namespace PersonalFinancer.Web.Areas.Identity.Pages.Account.Manage
 	{
 		private readonly UserManager<ApplicationUser> userManager;
 		private readonly SignInManager<ApplicationUser> signInManager;
-		private readonly ILogger<TwoFactorAuthenticationModel> logger;
 
 		public TwoFactorAuthenticationModel(
 			UserManager<ApplicationUser> userManager, 
-			SignInManager<ApplicationUser> signInManager, 
-			ILogger<TwoFactorAuthenticationModel> logger)
+			SignInManager<ApplicationUser> signInManager)
 		{
 			this.userManager = userManager;
 			this.signInManager = signInManager;
-			this.logger = logger;
 		}
 
 		public bool HasAuthenticator { get; set; }
@@ -37,6 +34,7 @@ namespace PersonalFinancer.Web.Areas.Identity.Pages.Account.Manage
 		public async Task<IActionResult> OnGetAsync()
 		{
 			var user = await userManager.GetUserAsync(User);
+
 			if (user == null)
 			{
 				return NotFound($"Unable to load user with ID '{userManager.GetUserId(User)}'.");
@@ -53,6 +51,7 @@ namespace PersonalFinancer.Web.Areas.Identity.Pages.Account.Manage
 		public async Task<IActionResult> OnPostAsync()
 		{
 			var user = await userManager.GetUserAsync(User);
+
 			if (user == null)
 			{
 				return NotFound($"Unable to load user with ID '{userManager.GetUserId(User)}'.");
@@ -60,6 +59,7 @@ namespace PersonalFinancer.Web.Areas.Identity.Pages.Account.Manage
 
 			await signInManager.ForgetTwoFactorClientAsync();
 			StatusMessage = "The current browser has been forgotten. When you login again from this browser you will be prompted for your 2fa code.";
+			
 			return RedirectToPage();
 		}
 	}
