+++ PersonalFinancer.Web/Controllers/Api/CategoriesApiController.cs
@@ -3,11 +3,11 @@ using PersonalFinancer.Data.Models;
 using PersonalFinancer.Services.ApiService;
 using PersonalFinancer.Services.ApiService.Models;
 using PersonalFinancer.Services.Shared.Models;
-using PersonalFinancer.Web.Infrastructure;
+using PersonalFinancer.Web.Infrastructure.Extensions;
 
 namespace PersonalFinancer.Web.Controllers.Api
 {
-	[Route("api/categories")]
+    [Route("api/categories")]
 	[ApiController]
 	public class CategoriesApiController : BaseApiController
 	{
