+++ PersonalFinancer.Web/Views/Shared/_Layout.cshtml
@@ -1,4 +1,4 @@
-﻿@using PersonalFinancer.Web.Infrastructure;
+﻿@using PersonalFinancer.Web.Infrastructure.Extensions;
 
 <!DOCTYPE html>
 <html lang="en">
