+++ PersonalFinancer.Web/Views/Shared/_LoginPartial.cshtml
@@ -1,7 +1,7 @@
 ﻿@using Microsoft.AspNetCore.Identity
 @using PersonalFinancer.Data.Models;
 @using PersonalFinancer.Services.User;
-@using PersonalFinancer.Web.Infrastructure;
+@using PersonalFinancer.Web.Infrastructure.Extensions;
 @inject IUsersService userService
 
 <ul class="navbar-nav">
